+++ lib/oneview-sdk/resource/api200/firmware_bundle.rb
@@ -14,7 +14,6 @@ module OneviewSDK
     # Firmware bundle resource implementation
     class FirmwareBundle
       BASE_URI = '/rest/firmware-bundles'.freeze
-      BOUNDARY = '----011000010111000001101001'.freeze
 
       # Uploads a firmware bundle file
       # @param [OneviewSDK::Client] client The client object for the OneView appliance
@@ -23,15 +22,28 @@ module OneviewSDK
       def self.add(client, file_path)
         raise NotFound, "ERROR: File '#{file_path}' not found!" unless File.file?(file_path)
         options = {}
-        options['Content-Type'] = "multipart/form-data; boundary=#{BOUNDARY}"
+        options['Content-Type'] = 'multipart/form-data'
+        options['X-Api-Version'] = client.api_version.to_s
+        options['auth'] = client.token
         options['uploadfilename'] = File.basename(file_path)
-        options['body'] = "--#{BOUNDARY}\r\n"
-        options['body'] << "Content-Disposition: form-data; name=\"file\"; filename=\"#{File.basename(file_path)}\"\r\n"
-        options['body'] << "Content-Type: application/octet-stream; Content-Transfer-Encoding: binary\r\n\r\n"
-        options['body'] << "#{IO.binread(file_path)}\r\n--#{BOUNDARY}--"
-        response = client.rest_post(BASE_URI, options)
-        data = client.response_handler(response)
-        OneviewSDK::FirmwareDriver.new(client, data)
+        url = URI.parse(URI.escape("#{client.url}#{BASE_URI}"))
+
+        File.open(file_path) do |file|
+          req = Net::HTTP::Post::Multipart.new(
+            url.path,
+            { 'file' => UploadIO.new(file, 'application/octet-stream', File.basename(file_path)) },
+            options
+          )
+
+          http_request = Net::HTTP.new(url.host, url.port)
+          http_request.use_ssl = true
+          http_request.verify_mode = OpenSSL::SSL::VERIFY_NONE
+          http_request.start do |http|
+            response = http.request(req)
+            data = client.response_handler(response)
+            return OneviewSDK::FirmwareDriver.new(client, data)
+          end
+        end
       end
     end
   end