+++ spec/unit/resource/api300/thunderbird/firmware_bundle_spec.rb
@@ -1,4 +1,5 @@
 require 'spec_helper'
+require 'net/http/post/multipart'
 
 RSpec.describe OneviewSDK::API300::Thunderbird::FirmwareBundle do
   include_context 'shared context'
@@ -13,17 +14,13 @@ RSpec.describe OneviewSDK::API300::Thunderbird::FirmwareBundle do
     end
 
     it 'returns a FirmwareDriver resource' do
-      expect(@client_300).to receive(:rest_post).with(
-        OneviewSDK::API300::Thunderbird::FirmwareBundle::BASE_URI,
-        hash_including(
-          'body' => %r{Content-Type: application/octet-stream; Content-Transfer-Encoding: binary\r\n\r\nFAKE FILE CONTENT\r\n--},
-          'uploadfilename' => 'file.tar',
-          'Content-Type' => "multipart/form-data; boundary=#{OneviewSDK::API300::Thunderbird::FirmwareBundle::BOUNDARY}"
-        )
-      ).and_return(FakeResponse.new({}))
-      expect(File).to receive('file?').and_return(true)
-      allow(IO).to receive(:binread).and_return('FAKE FILE CONTENT')
-      OneviewSDK::API300::Thunderbird::FirmwareBundle.add(@client_300, 'file.tar')
+      allow_any_instance_of(Net::HTTP).to receive(:request).and_return(FakeResponse.new({}, 200))
+      allow_any_instance_of(Net::HTTP).to receive(:connect).and_return(true)
+      allow(@client_300).to receive(:response_handler).and_return(uri: '/rest/firmware-drivers/f1')
+      allow(File).to receive(:file?).and_return(true)
+      allow(File).to receive(:open).with('file.tar').and_yield('FAKE FILE CONTENT')
+      allow(UploadIO).to receive(:new).and_return('FAKE FILE CONTENT')
+      expect(described_class.add(@client_300, 'file.tar').class).to eq(OneviewSDK::FirmwareDriver)
     end
   end
 end